requirements.txt
fastapi~=0.111.0
psycopg2-binary
alembic~=1.13.1
python-dotenv~=1.0.1
SQLAlchemy~=2.0.31
torch~=2.3.1
transformers~=4.41.2
pillow~=10.3.0
